+++ src/main/java/com/codeborne/selenide/SelenideElement.java
@@ -2,16 +2,11 @@ package com.codeborne.selenide;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.NoSuchElementException;
+import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.internal.FindsByClassName;
-import org.openqa.selenium.internal.FindsByCssSelector;
-import org.openqa.selenium.internal.FindsById;
-import org.openqa.selenium.internal.FindsByLinkText;
-import org.openqa.selenium.internal.FindsByName;
-import org.openqa.selenium.internal.FindsByTagName;
-import org.openqa.selenium.internal.FindsByXPath;
-import org.openqa.selenium.internal.WrapsDriver;
+import org.openqa.selenium.WrapsDriver;
 import org.openqa.selenium.internal.WrapsElement;
+import org.openqa.selenium.interactions.internal.Locatable;
 
 import java.awt.image.BufferedImage;
 import java.io.File;
@@ -21,10 +16,7 @@ import java.io.FileNotFoundException;
  * Wrapper around {@link WebElement} with additional methods like
  * {@link #shouldBe(Condition...)} and {@link #shouldHave(Condition...)}
  */
-public interface SelenideElement extends WebElement, FindsByLinkText, FindsById, FindsByName,
-    FindsByTagName, FindsByClassName, FindsByCssSelector,
-    FindsByXPath, WrapsDriver, WrapsElement, org.openqa.selenium.interactions.internal.Locatable {
-
+public interface SelenideElement extends WebElement, WrapsDriver, WrapsElement, Locatable, TakesScreenshot {
   /**
    *
    * @Deprecated this is a very legacy method with strange implementation. Don't use it. Use #click on links too